requirements.txt
Pillow>=10.0.0
imageio>=2.31.0
tkinterdnd2>=0.3.0
numpy>=1.24.0